+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme.NoActionBar">
 
-        <activity android:name=".SignInActivity">
+        <activity android:name=".authenticate.SignInActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -52,11 +52,11 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".NavigationActivity"
+            android:name=".navigation.NavigationActivity"
             android:label="@string/title_activity_navigation"
             android:theme="@style/AppTheme.NoActionBar"/>
-        <activity android:name=".SearchingActivity"/>
-        <activity android:name=".SettingActivity"/>
+        <activity android:name=".search.SearchingActivity"/>
+        <activity android:name=".setting.SettingActivity"/>
     </application>
 
 </manifest>
