+++ app/src/main/java/edu/uw/tacoma/team8/drinkndial/authenticate/LoginFragment.java
@@ -1,4 +1,4 @@
-package edu.uw.tacoma.team8.drinkndial;
+package edu.uw.tacoma.team8.drinkndial.authenticate;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -9,6 +9,8 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import edu.uw.tacoma.team8.drinkndial.R;
 
 
 /**
