--- app/src/main/java/edu/uw/tacoma/team8/drinkndial/model/Location.java
@@ -1,10 +1,0 @@
-package edu.uw.tacoma.team8.drinkndial.model;
-
-/**
- * Created by leejieun on 2/5/17.
- */
-
-public class Location {
-
-
-}
