+++ app/src/main/java/edu/uw/tacoma/team8/drinkndial/search/MyRecentLocationsRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package edu.uw.tacoma.team8.drinkndial;
+package edu.uw.tacoma.team8.drinkndial.search;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -8,7 +8,8 @@ import android.widget.TextView;
 
 import java.util.List;
 
-import edu.uw.tacoma.team8.drinkndial.fragments.RecentLocationsFragment.OnListFragmentInteractionListener;
+import edu.uw.tacoma.team8.drinkndial.R;
+import edu.uw.tacoma.team8.drinkndial.search.RecentLocationsFragment.OnListFragmentInteractionListener;
 import edu.uw.tacoma.team8.drinkndial.location.LocationContent;
 import edu.uw.tacoma.team8.drinkndial.location.LocationContent.LocationItem;
