+++ app/src/main/res/layout/activity_searching.xml
@@ -8,6 +8,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="edu.uw.tacoma.team8.drinkndial.SearchingActivity">
+    tools:context="edu.uw.tacoma.team8.drinkndial.search.SearchingActivity">
 
 </RelativeLayout>
