+++ app/src/main/res/layout/activity_setting.xml
@@ -3,6 +3,6 @@
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.uw.tacoma.team8.drinkndial.SettingActivity">
+    tools:context="edu.uw.tacoma.team8.drinkndial.setting.SettingActivity">
 
 </FrameLayout>
