+++ app/src/main/res/layout/app_bar_navigation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="edu.uw.tacoma.team8.drinkndial.NavigationActivity">
+    tools:context="edu.uw.tacoma.team8.drinkndial.navigation.NavigationActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
