+++ app/src/main/res/layout/content_navigation.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="edu.uw.tacoma.team8.drinkndial.NavigationActivity"
+    tools:context="edu.uw.tacoma.team8.drinkndial.navigation.NavigationActivity"
     tools:showIn="@layout/app_bar_navigation">
 
 -->
