+++ app/src/main/res/layout/fragment_recentlocations_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context="edu.uw.tacoma.team8.drinkndial.fragments.RecentLocationsFragment"
+    tools:context="edu.uw.tacoma.team8.drinkndial.search.RecentLocationsFragment"
     tools:listitem="@layout/fragment_recentlocations" />
