+++ app/src/main/res/layout/fragment_search_input.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.uw.tacoma.team8.drinkndial.fragments.SearchInputFragment">
+    tools:context="edu.uw.tacoma.team8.drinkndial.search.SearchInputFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
