+++ app/src/main/res/layout/fragment_setting_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="edu.uw.tacoma.team8.drinkndial.SettingActivity">
+    tools:context="edu.uw.tacoma.team8.drinkndial.setting.SettingActivity">
 
     <TextView
         android:layout_width="match_parent"
