+++ app/src/main/res/layout/fragment_settings_favorite_location.xml
@@ -8,7 +8,7 @@
     android:layout_marginLeft="15dp"
     android:layout_marginTop="15dp"
     android:layout_marginEnd="15dp"
-    tools:context="edu.uw.tacoma.team8.drinkndial.fragments.SettingsFavoriteLocation">
+    tools:context="edu.uw.tacoma.team8.drinkndial.setting.SettingsFavoriteLocation">
 
 
         <TextView
