+++ app/src/main/res/layout/fragment_settings_user.xml
@@ -9,7 +9,7 @@
     android:layout_marginLeft="15dp"
     android:layout_marginTop="15dp"
     android:layout_marginEnd="15dp"
-    tools:context="edu.uw.tacoma.team8.drinkndial.fragments.SettingsFragment">
+    tools:context="edu.uw.tacoma.team8.drinkndial.setting.SettingsFragment">
 
         <TextView
             style="@style/TitleTextStyle"
