+++ app/src/main/res/layout/fragment_trips.xml
@@ -4,7 +4,7 @@
     android:id="@+id/fragment_trips"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.uw.tacoma.team8.drinkndial.fragments.TripsFragment">
+    tools:context="edu.uw.tacoma.team8.drinkndial.navigation.TripsFragment">
 
 
     <TextView
