+++ app/src/main/res/layout/nav_header_navigation.xml
@@ -27,6 +27,7 @@
 
 
     <TextView
+        android:id="@+id/user_name_navigation"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="bottom"
@@ -37,7 +38,6 @@
         android:paddingEnd="20dp"
         android:layout_column="1"
         android:layout_row="0"
-        android:text="First Name, Last Name"
         android:textAppearance="@style/TextAppearance.AppCompat.Body1"
         android:textColor="@color/drinkWhite"/>
